+++ packages/ponder/src/index.ts
@@ -45,7 +45,7 @@ for (const name of ["RealityETH_v3_2", "RealityETH_v3_0"] as const) {
         user,
         questionText: stripNullBytes(question_text),
         createdBlock: event.block.number,
-        createdLogIndex: event.log.logIndex,
+        createdLogIndex: BigInt(event.log.logIndex),
         createdTxHash: event.log.transactionHash,
         createdTimestamp: event.block.timestamp,
       })
@@ -107,7 +107,7 @@ for (const name of ["RealityETH_v3_2", "RealityETH_v3_0"] as const) {
         arbitrationOccurred: false,
         scheduledFinalizationTimestamp: BigInt(opening_ts) + BigInt(timeout),
         createdBlock: event.block.number,
-        createdLogIndex: event.log.logIndex,
+        createdLogIndex: BigInt(event.log.logIndex),
         createdTxHash: event.log.transactionHash,
         createdTimestamp: event.block.timestamp,
         updatedBlock: event.block.number,
@@ -137,7 +137,7 @@ for (const name of ["RealityETH_v3_2", "RealityETH_v3_0"] as const) {
         isUnrevealed: is_commitment,
         timestamp: ts,
         createdBlock: event.block.number,
-        createdLogIndex: event.log.logIndex,
+        createdLogIndex: BigInt(event.log.logIndex),
         createdTxHash: event.log.transactionHash,
       })
       .onConflictDoNothing();
